Rename TestGrouperApiService to TestGrouperApiServiceTutorial, add more implementation to tests
Delete unused imports and private fields

Refactor subjectLookupTest
